requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
pydantic>=2.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
